[MINOR] fix some tests to pass CI
