+++ hudi-spark/src/test/scala/org/apache/hudi/functional/TestCOWDataSource.scala
@@ -17,34 +17,24 @@
 
 package org.apache.hudi.functional
 
-import org.apache.hadoop.fs.{FileSystem, Path}
-import org.apache.hudi.common.fs.FSUtils
-import org.apache.hudi.common.table.HoodieTableMetaClient
-import org.apache.hudi.common.testutils.HoodieTestDataGenerator
 import org.apache.hudi.common.testutils.RawTripTestPayload.recordsToStrings
 import org.apache.hudi.config.HoodieWriteConfig
-import org.apache.hudi.exception.TableNotFoundException
+import org.apache.hudi.testutils.HoodieClientTestBase
 import org.apache.hudi.{DataSourceReadOptions, DataSourceWriteOptions, HoodieDataSourceHelpers}
 import org.apache.log4j.LogManager
 import org.apache.spark.sql._
 import org.apache.spark.sql.functions.col
-import org.apache.spark.sql.streaming.{OutputMode, ProcessingTime}
 import org.junit.jupiter.api.Assertions.{assertEquals, assertTrue}
-import org.junit.jupiter.api.io.TempDir
-import org.junit.jupiter.api.{BeforeEach, Test}
+import org.junit.jupiter.api.{AfterEach, BeforeEach, Test}
 
 import scala.collection.JavaConversions._
-import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.duration.Duration
-import scala.concurrent.{Await, Future}
 
 /**
  * Basic tests on the spark datasource for COW table.
  */
-class TestCOWDataSource {
+class TestCOWDataSource extends HoodieClientTestBase {
   private val log = LogManager.getLogger(getClass)
   var spark: SparkSession = null
-  var dataGen: HoodieTestDataGenerator = null
   val commonOpts = Map(
     "hoodie.insert.shuffle.parallelism" -> "4",
     "hoodie.upsert.shuffle.parallelism" -> "4",
@@ -53,24 +43,25 @@ class TestCOWDataSource {
     DataSourceWriteOptions.PRECOMBINE_FIELD_OPT_KEY -> "timestamp",
     HoodieWriteConfig.TABLE_NAME -> "hoodie_test"
   )
-  var basePath: String = null
-  var fs: FileSystem = null
 
-  @BeforeEach def initialize(@TempDir tempDir: java.nio.file.Path) {
-    spark = SparkSession.builder
-      .appName("Hoodie Datasource test")
-      .master("local[2]")
-      .config("spark.serializer", "org.apache.spark.serializer.KryoSerializer")
-      .getOrCreate
-    dataGen = new HoodieTestDataGenerator()
-    basePath = tempDir.toAbsolutePath.toString
-    fs = FSUtils.getFs(basePath, spark.sparkContext.hadoopConfiguration)
+  @BeforeEach override def setUp() {
+    initPath()
+    initSparkContexts()
+    spark = sqlContext.sparkSession
+    initTestDataGenerator()
+    initFileSystem()
+  }
+
+  @AfterEach override def tearDown() = {
+    cleanupSparkContexts()
+    cleanupTestDataGenerator()
+    cleanupFileSystem()
   }
 
   @Test def testShortNameStorage() {
     // Insert Operation
     val records = recordsToStrings(dataGen.generateInserts("000", 100)).toList
-    val inputDF: Dataset[Row] = spark.read.json(spark.sparkContext.parallelize(records, 2))
+    val inputDF = spark.read.json(spark.sparkContext.parallelize(records, 2))
     inputDF.write.format("hudi")
       .options(commonOpts)
       .option(DataSourceWriteOptions.OPERATION_OPT_KEY, DataSourceWriteOptions.INSERT_OPERATION_OPT_VAL)
@@ -83,7 +74,7 @@ class TestCOWDataSource {
   @Test def testCopyOnWriteStorage() {
     // Insert Operation
     val records1 = recordsToStrings(dataGen.generateInserts("000", 100)).toList
-    val inputDF1: Dataset[Row] = spark.read.json(spark.sparkContext.parallelize(records1, 2))
+    val inputDF1 = spark.read.json(spark.sparkContext.parallelize(records1, 2))
     inputDF1.write.format("org.apache.hudi")
       .options(commonOpts)
       .option(DataSourceWriteOptions.OPERATION_OPT_KEY, DataSourceWriteOptions.INSERT_OPERATION_OPT_VAL)
@@ -91,15 +82,15 @@ class TestCOWDataSource {
       .save(basePath)
 
     assertTrue(HoodieDataSourceHelpers.hasNewCommits(fs, basePath, "000"))
-    val commitInstantTime1: String = HoodieDataSourceHelpers.latestCommit(fs, basePath)
+    val commitInstantTime1 = HoodieDataSourceHelpers.latestCommit(fs, basePath)
 
     // Read RO View
     val hoodieROViewDF1 = spark.read.format("org.apache.hudi")
-      .load(basePath + "/*/*/*/*");
+      .load(basePath + "/*/*/*/*")
     assertEquals(100, hoodieROViewDF1.count())
 
     val records2 = recordsToStrings(dataGen.generateUpdates("001", 100)).toList
-    val inputDF2: Dataset[Row] = spark.read.json(spark.sparkContext.parallelize(records2, 2))
+    val inputDF2 = spark.read.json(spark.sparkContext.parallelize(records2, 2))
     val uniqueKeyCnt = inputDF2.select("_row_key").distinct().count()
 
     // Upsert Operation
@@ -108,30 +99,30 @@ class TestCOWDataSource {
       .mode(SaveMode.Append)
       .save(basePath)
 
-    val commitInstantTime2: String = HoodieDataSourceHelpers.latestCommit(fs, basePath)
+    val commitInstantTime2 = HoodieDataSourceHelpers.latestCommit(fs, basePath)
     assertEquals(2, HoodieDataSourceHelpers.listCommitsSince(fs, basePath, "000").size())
 
     // Read RO View
     val hoodieROViewDF2 = spark.read.format("org.apache.hudi")
-      .load(basePath + "/*/*/*/*");
+      .load(basePath + "/*/*/*/*")
     assertEquals(100, hoodieROViewDF2.count()) // still 100, since we only updated
 
     // Read Incremental View
     // we have 2 commits, try pulling the first commit (which is not the latest)
-    val firstCommit = HoodieDataSourceHelpers.listCommitsSince(fs, basePath, "000").get(0);
+    val firstCommit = HoodieDataSourceHelpers.listCommitsSince(fs, basePath, "000").get(0)
     val hoodieIncViewDF1 = spark.read.format("org.apache.hudi")
       .option(DataSourceReadOptions.QUERY_TYPE_OPT_KEY, DataSourceReadOptions.QUERY_TYPE_INCREMENTAL_OPT_VAL)
       .option(DataSourceReadOptions.BEGIN_INSTANTTIME_OPT_KEY, "000")
       .option(DataSourceReadOptions.END_INSTANTTIME_OPT_KEY, firstCommit)
-      .load(basePath);
+      .load(basePath)
     assertEquals(100, hoodieIncViewDF1.count()) // 100 initial inserts must be pulled
-    var countsPerCommit = hoodieIncViewDF1.groupBy("_hoodie_commit_time").count().collect();
+    var countsPerCommit = hoodieIncViewDF1.groupBy("_hoodie_commit_time").count().collect()
     assertEquals(1, countsPerCommit.length)
     assertEquals(firstCommit, countsPerCommit(0).get(0))
 
     // Upsert an empty dataFrame
     val emptyRecords = recordsToStrings(dataGen.generateUpdates("002", 0)).toList
-    val emptyDF: Dataset[Row] = spark.read.json(spark.sparkContext.parallelize(emptyRecords, 1))
+    val emptyDF = spark.read.json(spark.sparkContext.parallelize(emptyRecords, 1))
     emptyDF.write.format("org.apache.hudi")
       .options(commonOpts)
       .mode(SaveMode.Append)
@@ -141,10 +132,10 @@ class TestCOWDataSource {
     val hoodieIncViewDF2 = spark.read.format("org.apache.hudi")
       .option(DataSourceReadOptions.QUERY_TYPE_OPT_KEY, DataSourceReadOptions.QUERY_TYPE_INCREMENTAL_OPT_VAL)
       .option(DataSourceReadOptions.BEGIN_INSTANTTIME_OPT_KEY, commitInstantTime1)
-      .load(basePath);
+      .load(basePath)
 
     assertEquals(uniqueKeyCnt, hoodieIncViewDF2.count()) // 100 records must be pulled
-    countsPerCommit = hoodieIncViewDF2.groupBy("_hoodie_commit_time").count().collect();
+    countsPerCommit = hoodieIncViewDF2.groupBy("_hoodie_commit_time").count().collect()
     assertEquals(1, countsPerCommit.length)
     assertEquals(commitInstantTime2, countsPerCommit(0).get(0))
 
@@ -153,7 +144,7 @@ class TestCOWDataSource {
       .option(DataSourceReadOptions.QUERY_TYPE_OPT_KEY, DataSourceReadOptions.QUERY_TYPE_INCREMENTAL_OPT_VAL)
       .option(DataSourceReadOptions.BEGIN_INSTANTTIME_OPT_KEY, commitInstantTime1)
       .option(DataSourceReadOptions.INCR_PATH_GLOB_OPT_KEY, "/2016/*/*/*")
-      .load(basePath);
+      .load(basePath)
     assertEquals(hoodieIncViewDF2.filter(col("_hoodie_partition_path").contains("2016")).count(), hoodieIncViewDF3.count())
   }
 
@@ -169,7 +160,7 @@ class TestCOWDataSource {
     val inserts2Dup = dataGen.generateSameKeyInserts("002", inserts1.subList(0, insert2DupKeyCnt))
 
     val records1 = recordsToStrings(inserts1).toList
-    val inputDF1: Dataset[Row] = spark.read.json(spark.sparkContext.parallelize(records1, 2))
+    val inputDF1 = spark.read.json(spark.sparkContext.parallelize(records1, 2))
     inputDF1.write.format("org.apache.hudi")
       .options(commonOpts)
       .option(DataSourceWriteOptions.OPERATION_OPT_KEY, DataSourceWriteOptions.INSERT_OPERATION_OPT_VAL)
@@ -181,7 +172,7 @@ class TestCOWDataSource {
 
     val commitInstantTime1 = HoodieDataSourceHelpers.latestCommit(fs, basePath)
     val records2 = recordsToStrings(inserts2Dup ++ inserts2New).toList
-    val inputDF2: Dataset[Row] = spark.read.json(spark.sparkContext.parallelize(records2, 2))
+    val inputDF2 = spark.read.json(spark.sparkContext.parallelize(records2, 2))
     inputDF2.write.format("org.apache.hudi")
       .options(commonOpts)
       .option(DataSourceWriteOptions.INSERT_DROP_DUPS_OPT_KEY, "true")
@@ -196,121 +187,5 @@ class TestCOWDataSource {
       .option(DataSourceReadOptions.BEGIN_INSTANTTIME_OPT_KEY, commitInstantTime1)
       .load(basePath)
     assertEquals(hoodieIncViewDF2.count(), insert2NewKeyCnt)
-  }
-
-  @Test
-  def testStructuredStreaming(): Unit = {
-    fs.delete(new Path(basePath), true)
-    val sourcePath = basePath + "/source"
-    val destPath = basePath + "/dest"
-    fs.mkdirs(new Path(sourcePath))
-
-    // First chunk of data
-    val records1 = recordsToStrings(dataGen.generateInserts("000", 100)).toList
-    val inputDF1: Dataset[Row] = spark.read.json(spark.sparkContext.parallelize(records1, 2))
-
-    // Second chunk of data
-    val records2 = recordsToStrings(dataGen.generateUpdates("001", 100)).toList
-    val inputDF2: Dataset[Row] = spark.read.json(spark.sparkContext.parallelize(records2, 2))
-    val uniqueKeyCnt = inputDF2.select("_row_key").distinct().count()
-
-    // define the source of streaming
-    val streamingInput =
-      spark.readStream
-      .schema(inputDF1.schema)
-      .json(sourcePath)
-
-    val f1 = Future {
-      println("streaming starting")
-    //'writeStream' can be called only on streaming Dataset/DataFrame
-      streamingInput
-        .writeStream
-        .format("org.apache.hudi")
-        .options(commonOpts)
-        .trigger(new ProcessingTime(100))
-        .option("checkpointLocation", basePath + "/checkpoint")
-        .outputMode(OutputMode.Append)
-        .start(destPath)
-        .awaitTermination(10000)
-      println("streaming ends")
-    }
-
-    val f2 = Future {
-      inputDF1.write.mode(SaveMode.Append).json(sourcePath)
-      // wait for spark streaming to process one microbatch
-      val currNumCommits = waitTillAtleastNCommits(fs, destPath, 1, 120, 5);
-      assertTrue(HoodieDataSourceHelpers.hasNewCommits(fs, destPath, "000"))
-      val commitInstantTime1: String = HoodieDataSourceHelpers.latestCommit(fs, destPath)
-      // Read RO View
-      val hoodieROViewDF1 = spark.read.format("org.apache.hudi")
-        .load(destPath + "/*/*/*/*")
-      assert(hoodieROViewDF1.count() == 100)
-
-      inputDF2.write.mode(SaveMode.Append).json(sourcePath)
-      // wait for spark streaming to process one microbatch
-      waitTillAtleastNCommits(fs, destPath, currNumCommits + 1, 120, 5);
-      val commitInstantTime2: String = HoodieDataSourceHelpers.latestCommit(fs, destPath)
-      assertEquals(2, HoodieDataSourceHelpers.listCommitsSince(fs, destPath, "000").size())
-      // Read RO View
-      val hoodieROViewDF2 = spark.read.format("org.apache.hudi")
-        .load(destPath + "/*/*/*/*")
-      assertEquals(100, hoodieROViewDF2.count()) // still 100, since we only updated
-
-
-      // Read Incremental View
-      // we have 2 commits, try pulling the first commit (which is not the latest)
-      val firstCommit = HoodieDataSourceHelpers.listCommitsSince(fs, destPath, "000").get(0)
-      val hoodieIncViewDF1 = spark.read.format("org.apache.hudi")
-        .option(DataSourceReadOptions.QUERY_TYPE_OPT_KEY, DataSourceReadOptions.QUERY_TYPE_INCREMENTAL_OPT_VAL)
-        .option(DataSourceReadOptions.BEGIN_INSTANTTIME_OPT_KEY, "000")
-        .option(DataSourceReadOptions.END_INSTANTTIME_OPT_KEY, firstCommit)
-        .load(destPath)
-      assertEquals(100, hoodieIncViewDF1.count())
-      // 100 initial inserts must be pulled
-      var countsPerCommit = hoodieIncViewDF1.groupBy("_hoodie_commit_time").count().collect()
-      assertEquals(1, countsPerCommit.length)
-      assertEquals(firstCommit, countsPerCommit(0).get(0))
-
-      // pull the latest commit
-      val hoodieIncViewDF2 = spark.read.format("org.apache.hudi")
-        .option(DataSourceReadOptions.QUERY_TYPE_OPT_KEY, DataSourceReadOptions.QUERY_TYPE_INCREMENTAL_OPT_VAL)
-        .option(DataSourceReadOptions.BEGIN_INSTANTTIME_OPT_KEY, commitInstantTime1)
-        .load(destPath)
-
-      assertEquals(uniqueKeyCnt, hoodieIncViewDF2.count()) // 100 records must be pulled
-      countsPerCommit = hoodieIncViewDF2.groupBy("_hoodie_commit_time").count().collect()
-      assertEquals(1, countsPerCommit.length)
-      assertEquals(commitInstantTime2, countsPerCommit(0).get(0))
-    }
-    Await.result(Future.sequence(Seq(f1, f2)), Duration.Inf)
-  }
-
-  @throws[InterruptedException]
-  private def waitTillAtleastNCommits(fs: FileSystem, tablePath: String,
-                               numCommits: Int, timeoutSecs: Int, sleepSecsAfterEachRun: Int): Int = {
-    val beginTime = System.currentTimeMillis
-    var currTime = beginTime
-    val timeoutMsecs = timeoutSecs * 1000
-    var numInstants = 0
-    var success: Boolean = false
-    while ({!success && (currTime - beginTime) < timeoutMsecs}) try {
-      val timeline = HoodieDataSourceHelpers.allCompletedCommitsCompactions(fs, tablePath)
-      log.info("Timeline :" + timeline.getInstants.toArray)
-      if (timeline.countInstants >= numCommits) {
-        numInstants = timeline.countInstants
-        success = true
-      }
-      val metaClient = new HoodieTableMetaClient(fs.getConf, tablePath, true)
-    } catch {
-      case te: TableNotFoundException =>
-        log.info("Got table not found exception. Retrying")
-    } finally {
-      Thread.sleep(sleepSecsAfterEachRun * 1000)
-      currTime = System.currentTimeMillis
-    }
-    if (!success) {
-      throw new IllegalStateException("Timed-out waiting for " + numCommits + " commits to appear in " + tablePath)
-    }
-    numInstants
   }
 }
